refactor: :bulb:  merge settings
