--- flow/flows/abstract_demands.py
@@ -1,5 +1,0 @@
-from dataclasses import dataclass
-
-@dataclass
-class AbstractDemands:
-    pass
